+++ modules/base/index.esm.mjs
@@ -4,6 +4,8 @@
 
 export { default as GSLog } from "./GSLog.mjs";
 export { default as GSID } from "./GSID.mjs";
+export { default as GSAbortController } from "./GSAbortController.mjs";
+export { default as GSPromise } from "./GSPromise.mjs";
 export { default as GSFunction } from "./GSFunction.mjs";
 export { default as GSEvents } from "./GSEvents.mjs";
 export { default as GSEvent } from "./GSEvent.mjs";
@@ -27,3 +29,6 @@ export { default as GSi18n } from "./GSi18n.mjs";
 export { default as GSQueryCSS } from './GSQueryCSS.mjs';
 export { default as GSQuery } from './GSQuery.mjs';
 export { default as GSTouch } from './GSTouch.mjs';
+
+export { default as GSReadWrite} from './GSReadWrite.mjs';
+export { default as GSReadWriteRegistry} from './GSReadWriteRegistry.mjs';